Moved PLACE and ORDER to a separate file commands.c
